requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
Django==3.0.7
django-filter==2.3.0
django-filters==0.2.1
djangorestframework==3.11.0
future==0.18.2
idna==2.9
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.8.2
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.4
